Crash Handler UX

+++ SystemManagement/frmCrashReport.cs
@@ -39,8 +39,12 @@ namespace SystemManagement
 
         private async void cmdDelete_Click(object sender, EventArgs e)
         {
-            Enabled = false;
+            if(MessageBox.Show("Are you sure you want to delete this Crash Report?", "Delete Crash Report", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation) != DialogResult.Yes)
+            {
+                return;
+            }
 
+            Enabled = false;
             DeleteData delete = new DeleteData(DataAccess.APIs.SystemManagement, "Crash/Delete");
             delete.QueryString.Add("id", CrashReportID.ToString());
             await delete.Execute();
